# routes/swagger.yaml
openapi: 3.0.0
info:
  title: Commerce
  contact: {}
  version: '1.0'
servers:
  - url: https://node-course-e-commerce-8r2s.onrender.com/api/v1
    variables: {}
paths:
...
  /orders:
    post:
      tags:
        - Order
      summary: Create Order
      operationId: CreateOrder

      parameters:
        - name: id
          in: path
          description: ''
          required: true
          style: simple
          schema:
            type: string
            example: 648329d8913fcaed6533b271
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/CreateOrderRequest'
                - example:
                    product:
                      _id: 648329d8913fcaed6533b271
                      name: albany sectional
                      price: 109999
                      description: Cloud bread VHS hell of banjo bicycle rights jianbing umami mumblecore etsy 8-bit pok pok +1 wolf. Vexillologist yr dreamcatcher waistcoat, authentic chillwave trust fund. Viral typewriter fingerstache pinterest pork belly narwhal. Schlitz venmo everyday carry kitsch pitchfork chillwave iPhone taiyaki trust fund hashtag kinfolk microdosing gochujang live-edge
                      image: /uploads/example.jpeg
                      category: kitchen
                      company: liddy
                      colors:
                        - '#000'
                        - '#ffb900'
                      featured: false
                      freeShipping: false
                      inventory: 15
                      averageRating: 2
                      numOfReviews: 2
                      user: 64831f7512ffe94a561f4035
                      createdAt: 2023-06-09T13:32:08.232Z
                      updatedAt: 2023-06-12T08:57:44.534Z
                      __v: 0
                      reviews:
                        - _id: 6486d853a754d25fef672615
                          rating: 3
                          title: bad product
                          comment: very, very bad product by anna
                          user: 64831f7512ffe94a561f4035
                          product: 648329d8913fcaed6533b271
                          createdAt: 2023-06-12T08:33:23.516Z
                          updatedAt: 2023-06-12T08:57:44.459Z
                          __v: 0
                        - _id: 6486d8a2a754d25fef67261d
                          rating: 1
                          title: bad  product
                          comment: very, very bad product
                          user: 64831f8f12ffe94a561f403d
                          product: 648329d8913fcaed6533b271
                          createdAt: 2023-06-12T08:34:42.273Z
                          updatedAt: 2023-06-12T08:34:42.273Z
                          __v: 0
                      id: 648329d8913fcaed6533b271
            example:
              product:
                _id: 648329d8913fcaed6533b271
                name: albany sectional
                price: 109999
                description: Cloud bread VHS hell of banjo bicycle rights jianbing umami mumblecore etsy 8-bit pok pok +1 wolf. Vexillologist yr dreamcatcher waistcoat, authentic chillwave trust fund. Viral typewriter fingerstache pinterest pork belly narwhal. Schlitz venmo everyday carry kitsch pitchfork chillwave iPhone taiyaki trust fund hashtag kinfolk microdosing gochujang live-edge
                image: /uploads/example.jpeg
                category: kitchen
                company: liddy
                colors:
                  - '#000'
                  - '#ffb900'
                featured: false
                freeShipping: false
                inventory: 15
                averageRating: 2
                numOfReviews: 2
                user: 64831f7512ffe94a561f4035
                createdAt: 2023-06-09T13:32:08.232Z
                updatedAt: 2023-06-12T08:57:44.534Z
                __v: 0
                reviews:
                  - _id: 6486d853a754d25fef672615
                    rating: 3
                    title: bad product
                    comment: very, very bad product by anna
                    user: 64831f7512ffe94a561f4035
                    product: 648329d8913fcaed6533b271
                    createdAt: 2023-06-12T08:33:23.516Z
                    updatedAt: 2023-06-12T08:57:44.459Z
                    __v: 0
                  - _id: 6486d8a2a754d25fef67261d
                    rating: 1
                    title: bad  product
                    comment: very, very bad product
                    user: 64831f8f12ffe94a561f403d
                    product: 648329d8913fcaed6533b271
                    createdAt: 2023-06-12T08:34:42.273Z
                    updatedAt: 2023-06-12T08:34:42.273Z
                    __v: 0
                id: 648329d8913fcaed6533b271
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
    get:
      tags:
        - Order
      summary: Get All Orders
      operationId: GetAllOrders

      parameters:
        - name: id
          in: path
          description: ''
          required: true
          style: simple
          schema:
            type: string
            example: 648329d8913fcaed6533b271
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
  /orders/6487153f534200366cfc6041:
    get:
      tags:
        - Order
      summary: Get Single Order
      operationId: GetSingleOrder

      parameters:
        - name: id
          in: path
          description: ''
          required: true
          style: simple
          schema:
            type: string
            example: 648329d8913fcaed6533b271
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
  /orders/123:
    patch:
      tags:
        - Order
      summary: Update Order
      operationId: UpdateOrder

      parameters:
        - name: id
          in: path
          description: ''
          required: true
          style: simple
          schema:
            type: string
            example: 648329d8913fcaed6533b271
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
...
components:
  schemas:
...
    CreateOrderRequest:
      title: CreateOrderRequest
      required:
        - product
      type: object
      properties:
        product:
          allOf:
            - $ref: '#/components/schemas/Product'
            - {}
      example:
        product:
          _id: 648329d8913fcaed6533b271
          name: albany sectional
          price: 109999
          description: Cloud bread VHS hell of banjo bicycle rights jianbing umami mumblecore etsy 8-bit pok pok +1 wolf. Vexillologist yr dreamcatcher waistcoat, authentic chillwave trust fund. Viral typewriter fingerstache pinterest pork belly narwhal. Schlitz venmo everyday carry kitsch pitchfork chillwave iPhone taiyaki trust fund hashtag kinfolk microdosing gochujang live-edge
          image: /uploads/example.jpeg
          category: kitchen
          company: liddy
          colors:
            - '#000'
            - '#ffb900'
          featured: false
          freeShipping: false
          inventory: 15
          averageRating: 2
          numOfReviews: 2
          user: 64831f7512ffe94a561f4035
          createdAt: 2023-06-09T13:32:08.232Z
          updatedAt: 2023-06-12T08:57:44.534Z
          __v: 0
          reviews:
            - _id: 6486d853a754d25fef672615
              rating: 3
              title: bad product
              comment: very, very bad product by anna
              user: 64831f7512ffe94a561f4035
              product: 648329d8913fcaed6533b271
              createdAt: 2023-06-12T08:33:23.516Z
              updatedAt: 2023-06-12T08:57:44.459Z
              __v: 0
            - _id: 6486d8a2a754d25fef67261d
              rating: 1
              title: bad  product
              comment: very, very bad product
              user: 64831f8f12ffe94a561f403d
              product: 648329d8913fcaed6533b271
              createdAt: 2023-06-12T08:34:42.273Z
              updatedAt: 2023-06-12T08:34:42.273Z
              __v: 0
          id: 648329d8913fcaed6533b271
    Product:
      title: Product
      required:
        - _id
        - name
        - price
        - description
        - image
        - category
        - company
        - colors
        - featured
        - freeShipping
        - inventory
        - averageRating
        - numOfReviews
        - user
        - createdAt
        - updatedAt
        - __v
        - reviews
        - id
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        price:
          type: integer
          format: int32
        description:
          type: string
        image:
          type: string
        category:
          type: string
        company:
          type: string
        colors:
          type: array
          items:
            type: string
          description: ''
        featured:
          type: boolean
        freeShipping:
          type: boolean
        inventory:
          type: integer
          format: int32
        averageRating:
          type: integer
          format: int32
        numOfReviews:
          type: integer
          format: int32
        user:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        __v:
          type: integer
          format: int32
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/Review'
          description: ''
        id:
          type: string
      example:
        _id: 648329d8913fcaed6533b271
        name: albany sectional
        price: 109999
        description: Cloud bread VHS hell of banjo bicycle rights jianbing umami mumblecore etsy 8-bit pok pok +1 wolf. Vexillologist yr dreamcatcher waistcoat, authentic chillwave trust fund. Viral typewriter fingerstache pinterest pork belly narwhal. Schlitz venmo everyday carry kitsch pitchfork chillwave iPhone taiyaki trust fund hashtag kinfolk microdosing gochujang live-edge
        image: /uploads/example.jpeg
        category: kitchen
        company: liddy
        colors:
          - '#000'
          - '#ffb900'
        featured: false
        freeShipping: false
        inventory: 15
        averageRating: 2
        numOfReviews: 2
        user: 64831f7512ffe94a561f4035
        createdAt: 2023-06-09T13:32:08.232Z
        updatedAt: 2023-06-12T08:57:44.534Z
        __v: 0
        reviews:
          - _id: 6486d853a754d25fef672615
            rating: 3
            title: bad product
            comment: very, very bad product by anna
            user: 64831f7512ffe94a561f4035
            product: 648329d8913fcaed6533b271
            createdAt: 2023-06-12T08:33:23.516Z
            updatedAt: 2023-06-12T08:57:44.459Z
            __v: 0
          - _id: 6486d8a2a754d25fef67261d
            rating: 1
            title: bad  product
            comment: very, very bad product
            user: 64831f8f12ffe94a561f403d
            product: 648329d8913fcaed6533b271
            createdAt: 2023-06-12T08:34:42.273Z
            updatedAt: 2023-06-12T08:34:42.273Z
            __v: 0
        id: 648329d8913fcaed6533b271
    Review:
      title: Review
      required:
        - _id
        - rating
        - title
        - comment
        - user
        - product
        - createdAt
        - updatedAt
        - __v
      type: object
      properties:
        _id:
          type: string
        rating:
          type: integer
          format: int32
        title:
          type: string
        comment:
          type: string
        user:
          type: string
        product:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        __v:
          type: integer
          format: int32
      example:
        _id: 6486d853a754d25fef672615
        rating: 3
        title: bad product
        comment: very, very bad product by anna
        user: 64831f7512ffe94a561f4035
        product: 648329d8913fcaed6533b271
        createdAt: 2023-06-12T08:33:23.516Z
        updatedAt: 2023-06-12T08:57:44.459Z
        __v: 0
  securitySchemes:
    httpBearer:
      type: http
      scheme: bearer
security:
  - httpBearer: []
...